final task: debounce. Debounce does not working
